backend/gemini/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.12.3
pydantic-settings==2.7.0
asyncpg==0.29.0
redis==5.0.1
google-generativeai==0.3.2
python-multipart==0.0.6
